+++ app/src/main/java/com/monday8am/tweetmeck/di/modules.kt
@@ -10,12 +10,12 @@ import com.monday8am.tweetmeck.data.local.SharedPreferencesServiceImpl
 import com.monday8am.tweetmeck.data.local.TwitterDatabase
 import com.monday8am.tweetmeck.data.remote.TwitterClient
 import com.monday8am.tweetmeck.data.remote.TwitterClientImpl
-import com.monday8am.tweetmeck.home.HomeFragment
-import com.monday8am.tweetmeck.home.HomeViewModel
-import com.monday8am.tweetmeck.home.TimelinePoolProvider
-import com.monday8am.tweetmeck.login.AuthViewModel
-import com.monday8am.tweetmeck.tweet.TweetViewModel
-import com.monday8am.tweetmeck.user.UserViewModel
+import com.monday8am.tweetmeck.ui.home.HomeFragment
+import com.monday8am.tweetmeck.ui.home.HomeViewModel
+import com.monday8am.tweetmeck.ui.home.TimelinePoolProvider
+import com.monday8am.tweetmeck.ui.login.AuthViewModel
+import com.monday8am.tweetmeck.ui.tweet.TweetViewModel
+import com.monday8am.tweetmeck.ui.user.UserViewModel
 import org.koin.android.ext.koin.androidContext
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.core.qualifier.named
