+++ app/src/main/java/com/monday8am/tweetmeck/ui/tweet/TweetFragment.kt
@@ -1,4 +1,4 @@
-package com.monday8am.tweetmeck.user
+package com.monday8am.tweetmeck.ui.tweet
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,28 +7,26 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
-import com.monday8am.tweetmeck.databinding.UserFragmentBinding
-import com.monday8am.tweetmeck.login.AuthViewModel
+import com.monday8am.tweetmeck.databinding.TweetFragmentBinding
 import org.koin.androidx.viewmodel.ext.android.getViewModel
-import org.koin.androidx.viewmodel.ext.android.sharedViewModel
 import org.koin.core.parameter.parametersOf
 
-class UserFragment : Fragment() {
+class TweetFragment : Fragment() {
 
-    private val navArgs: UserFragmentArgs by navArgs()
-    private val authViewModel: AuthViewModel by sharedViewModel()
-    private lateinit var userViewModel: UserViewModel
+    private val navArgs: TweetFragmentArgs by navArgs()
+    private lateinit var binding: TweetFragmentBinding
+
+    private lateinit var viewModel: TweetViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        userViewModel = getViewModel { parametersOf(navArgs.userId) }
-        val binding = UserFragmentBinding.inflate(inflater, container, false).apply {
+        viewModel = getViewModel { parametersOf(navArgs.tweetId) }
+        binding = TweetFragmentBinding.inflate(inflater, container, false).apply {
             lifecycleOwner = viewLifecycleOwner
-            viewmodel = this@UserFragment.userViewModel
-            authViewModel = this@UserFragment.authViewModel
+            viewmodel = this@TweetFragment.viewModel
         }
 
         binding.toolbar.setNavigationOnClickListener {
