+++ app/src/main/java/com/monday8am/tweetmeck/ui/user/UserFragment.kt
@@ -1,4 +1,4 @@
-package com.monday8am.tweetmeck.tweet
+package com.monday8am.tweetmeck.ui.user
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,26 +7,28 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
-import com.monday8am.tweetmeck.databinding.TweetFragmentBinding
+import com.monday8am.tweetmeck.databinding.UserFragmentBinding
+import com.monday8am.tweetmeck.ui.login.AuthViewModel
 import org.koin.androidx.viewmodel.ext.android.getViewModel
+import org.koin.androidx.viewmodel.ext.android.sharedViewModel
 import org.koin.core.parameter.parametersOf
 
-class TweetFragment : Fragment() {
+class UserFragment : Fragment() {
 
-    private val navArgs: TweetFragmentArgs by navArgs()
-    private lateinit var binding: TweetFragmentBinding
-
-    private lateinit var viewModel: TweetViewModel
+    private val navArgs: UserFragmentArgs by navArgs()
+    private val authViewModel: AuthViewModel by sharedViewModel()
+    private lateinit var userViewModel: UserViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        viewModel = getViewModel { parametersOf(navArgs.tweetId) }
-        binding = TweetFragmentBinding.inflate(inflater, container, false).apply {
+        userViewModel = getViewModel { parametersOf(navArgs.userId) }
+        val binding = UserFragmentBinding.inflate(inflater, container, false).apply {
             lifecycleOwner = viewLifecycleOwner
-            viewmodel = this@TweetFragment.viewModel
+            viewmodel = this@UserFragment.userViewModel
+            authViewModel = this@UserFragment.authViewModel
         }
 
         binding.toolbar.setNavigationOnClickListener {
