+++ app/src/main/res/layout/fragment_auth_web_view.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.AuthWebViewFragment">
+    tools:context=".ui.login.AuthWebViewFragment">
     <WebView
         android:id="@+id/webView"
         android:layout_width="match_parent"
