+++ app/src/main/res/layout/home_fragment.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.monday8am.tweetmeck.home.HomeViewModel" />
+            type="com.monday8am.tweetmeck.ui.home.HomeViewModel" />
     </data>
 
     <com.monday8am.tweetmeck.widget.InsetDispatchingCoordinatorLayout
