+++ app/src/main/res/layout/user_fragment.xml
@@ -3,11 +3,11 @@
 <data>
     <variable
         name="viewmodel"
-        type="com.monday8am.tweetmeck.user.UserViewModel" />
+        type="com.monday8am.tweetmeck.ui.user.UserViewModel" />
 
     <variable
         name="authViewModel"
-        type="com.monday8am.tweetmeck.login.AuthViewModel" />
+        type="com.monday8am.tweetmeck.ui.login.AuthViewModel" />
 </data>
 <androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
@@ -18,7 +18,7 @@
     tools:ignore="MergeRootFrame"
     app:marginLeftSystemWindowInsets="@{true}"
     app:marginRightSystemWindowInsets="@{true}"
-    tools:context=".user.UserFragment">
+    tools:context=".ui.user.UserFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/tweet_detail_appbars"
